correction

+++ src/main/java/com/osm/conditioning/dto/BomLineDto.java
@@ -4,13 +4,12 @@ import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 @Data
 @Getter
 @Setter
 public class BomLineDto  {
     private UUID articleId;
-    private BigDecimal quantity;
+    private double quantity;
 
 }
